# docs/swagger.yaml
basePath: /
definitions:
...
  entity.Product:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for Go Cashier API.
  title: Go Cashier API
  version: '1.0'
paths:
...
  /api/products:
    get:
      consumes:
        - application/json
      description: Get a list of all products
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Product'
            type: array
        '500':
          description: Internal Server Error
          schema:
            type: string
      summary: List all products
      tags:
        - products
    post:
      consumes:
        - application/json
      description: Create a new product with the input payload
      parameters:
        - description: Product content
          in: body
          name: product
          required: true
          schema:
            $ref: '#/definitions/entity.Product'
      produces:
        - application/json
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/entity.Product'
        '400':
          description: Bad Request
          schema:
            type: string
        '500':
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new product
      tags:
        - products
...
swagger: '2.0'